refactor(core): separate interaction handler to it's own construct
