+++ src/shared/common-funtion-props.ts
